@@ -22,46 +22,47 @@
     "upgrade:latest": "pnpm upgrade --interactive --latest"
   },
   "dependencies": {
-    "@genql/cli": "^6.2.1",
-    "@hookform/resolvers": "^3.3.1",
+    "@genql/cli": "^6.3.1",
+    "@hookform/resolvers": "^3.3.4",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-collapsible": "^1.0.3",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-form": "^0.0.3",
     "@radix-ui/react-navigation-menu": "^1.1.4",
-    "@vercel/analytics": "^1.1.0",
-    "framer-motion": "^10.16.4",
-    "next": "14.0.2",
-    "next-intl": "3.0.0-beta.9",
+    "@vercel/analytics": "^1.1.2",
+    "framer-motion": "^11.0.3",
+    "next": "14.1.0",
+    "next-intl": "3.5.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.48.2",
+    "react-hook-form": "^7.50.0",
     "zod": "3.22.4"
   },
   "devDependencies": {
-    "@2digits/prettier-config": "^0.9.4",
-    "@storybook/addon-essentials": "^7.5.3",
-    "@storybook/addon-interactions": "^7.5.3",
-    "@storybook/addon-links": "^7.5.3",
-    "@storybook/blocks": "^7.5.3",
-    "@storybook/nextjs": "^7.5.3",
-    "@storybook/react": "^7.5.3",
+    "@storybook/addon-essentials": "^7.6.12",
+    "@storybook/addon-interactions": "^7.6.12",
+    "@storybook/addon-links": "^7.6.12",
+    "@storybook/blocks": "^7.6.12",
+    "@storybook/nextjs": "^7.6.12",
+    "@storybook/react": "^7.6.12",
     "@storybook/testing-library": "^0.2.2",
-    "@types/node": "20.9.0",
-    "@types/react": "18.2.37",
-    "@types/react-dom": "18.2.15",
-    "autoprefixer": "10.4.16",
-    "clsx": "^2.0.0",
-    "eslint": "8.53.0",
-    "eslint-config-next": "14.0.2",
-    "eslint-plugin-storybook": "^0.6.14",
-    "postcss": "8.4.31",
-    "prettier": "^2.8.8",
-    "storybook": "^7.5.3",
-    "tailwind-merge": "^2.0.0",
-    "tailwind-variants": "^0.1.18",
-    "tailwindcss": "3.3.5",
-    "typescript": "5.2.2"
+    "@types/node": "20.11.16",
+    "@types/react": "18.2.51",
+    "@types/react-dom": "18.2.18",
+    "autoprefixer": "10.4.17",
+    "clsx": "^2.1.0",
+    "eslint": "8.56.0",
+    "eslint-config-next": "14.1.0",
+    "eslint-plugin-storybook": "^0.6.15",
+    "eslint-plugin-unused-imports": "^3.0.0",
+    "postcss": "8.4.33",
+    "prettier": "^3.2.4",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "storybook": "^7.6.12",
+    "tailwind-merge": "^2.2.1",
+    "tailwind-variants": "^0.1.20",
+    "tailwindcss": "3.4.1",
+    "typescript": "5.3.3"
   },
-  "packageManager": "pnpm@8.10.2"
+  "packageManager": "pnpm@8.15.1"
 }
