@@ -1,14 +1,12 @@
 import { type PropsWithChildren } from 'react';
 
 import { Analytics } from '@vercel/analytics/react';
-import { NextIntlClientProvider } from 'next-intl';
-import { getTranslator } from 'next-intl/server';
-
-import { getTranslations } from '@/utils/translationUtils';
-
-import '@/styles/global.css';
 
 import { fonts } from '@/constants/fonts';
+import { getTranslations, unstable_setRequestLocale } from 'next-intl/server';
+import { locales } from '@/navigation';
+import { NextIntlClientProvider } from 'next-intl';
+import translations from '@/translations/en';
 
 export interface ParamsProps {
   params: { locale: string };
@@ -16,8 +14,12 @@ export interface ParamsProps {
 
 export interface RootLayoutProps extends PropsWithChildren<ParamsProps> {}
 
+export function generateStaticParams() {
+  return locales.map((locale) => ({ locale }));
+}
+
 export async function generateMetadata({ params: { locale } }: ParamsProps) {
-  const t = await getTranslator(locale, 'common.meta');
+  const t = await getTranslations({ locale, namespace: 'common.meta' });
 
   return {
     title: {
@@ -28,17 +30,17 @@ export async function generateMetadata({ params: { locale } }: ParamsProps) {
   };
 }
 
-export default async function RootLayout({ children, params: { locale } }: RootLayoutProps) {
-  const translations = await getTranslations(locale);
+export default async function LocaleLayout({ children, params: { locale } }: RootLayoutProps) {
+  unstable_setRequestLocale(locale);
 
   return (
     <html lang={locale} className={fonts}>
       <body className="flex min-h-screen flex-col">
         <NextIntlClientProvider locale={locale} messages={translations}>
           {children}
-
-          <Analytics />
         </NextIntlClientProvider>
+
+        <Analytics />
       </body>
     </html>
   );
