@@ -14,7 +14,7 @@ export function Item({ className, year, children }: ItemProps) {
         'pt-[6.25rem]': !!year,
       })}>
       {!!year && (
-        <span className="absolute top-0 z-10 truncate text-9xl font-black text-theme-brand-100">
+        <span className="absolute top-0 truncate text-9xl font-black text-theme-brand-100">
           {year}
         </span>
       )}
