@@ -24,18 +24,18 @@ export type FieldsSelection<SRC extends Anify<DST> | undefined, DST> = {
 }[DST extends Nil
   ? 'never'
   : DST extends false | 0
-  ? 'never'
-  : SRC extends Scalar
-  ? 'scalar'
-  : SRC extends any[]
-  ? 'array'
-  : SRC extends { __isUnion?: any }
-  ? 'union'
-  : DST extends { __scalar?: any }
-  ? '__scalar'
-  : DST extends {}
-  ? 'object'
-  : 'never'];
+    ? 'never'
+    : SRC extends Scalar
+      ? 'scalar'
+      : SRC extends any[]
+        ? 'array'
+        : SRC extends { __isUnion?: any }
+          ? 'union'
+          : DST extends { __scalar?: any }
+            ? '__scalar'
+            : DST extends {}
+              ? 'object'
+              : 'never'];
 
 type HandleObject<SRC extends Anify<DST>, DST> = SRC extends Nil
   ? never
@@ -65,12 +65,12 @@ type Handle__scalar<SRC extends Anify<DST>, DST> = SRC extends Nil
         [Key in keyof SRC]: SRC[Key] extends Nil
           ? never
           : Key extends FieldsToRemove
-          ? never
-          : SRC[Key] extends Scalar
-          ? Key
-          : Key extends keyof DST
-          ? Key
-          : never;
+            ? never
+            : SRC[Key] extends Scalar
+              ? Key
+              : Key extends keyof DST
+                ? Key
+                : never;
       }[keyof SRC]
     >;
 
